Update to version 3.1.1 beta

+++ apps/block_scout_web/lib/block_scout_web/schema/types.ex
@@ -137,7 +137,7 @@ defmodule BlockScoutWeb.Schema.Types do
     field(:r, :decimal)
     field(:s, :decimal)
     field(:status, :status)
-    field(:v, :integer)
+    field(:v, :decimal)
     field(:value, :wei)
     field(:from_address_hash, :address_hash)
     field(:to_address_hash, :address_hash)
